Correction bug génération de la base de données et verrouillage des pages avec Authenticator

+++ src/Controller/BookDelete.php
@@ -1,18 +1,24 @@
 <?php
 
 use Entity\Book;
+use Security\Authenticator;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
 /**
  * @var Twig\Environment $twig
  * @var Doctrine\ORM\EntityManager $entityManager
  * @var int $id
  */
 
+Authenticator::init();
 
-$bookRepository = $entityManager->getRepository(Book::class);
+if (Authenticator::is_authenticated()) {
+    $bookRepository = $entityManager->getRepository(Book::class);
+    $book = $bookRepository->find($id);
+    $entityManager->remove($book);
+    $entityManager->flush();
 
-$book = $bookRepository->find($id);
-$entityManager->remove($book);
-$entityManager->flush();
-
-return new \Symfony\Component\HttpFoundation\RedirectResponse('/book');
-?>
+    return new \Symfony\Component\HttpFoundation\RedirectResponse('/book');
+} else {
+    return new RedirectResponse(Authenticator::urlNotLogged());
+}
